@@ -5,14 +5,14 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.5.6")
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springbootVersion}")
     }
 }
 
 plugins {
     id 'com.palantir.docker' version "0.34.0"
     id "com.github.jk1.dependency-license-report" version "1.17"
-    id "com.github.hierynomus.license" version "0.15.0"
+    id "com.github.hierynomus.license" version "0.16.1"
     id "org.sonarqube" version "3.5.0.2730"
     id 'net.researchgate.release' version '3.0.2'
 }
@@ -26,7 +26,7 @@ release {
 
 allprojects {
     // TODO remove
-    apply plugin: 'maven'
+    apply plugin: 'maven-publish'
     apply plugin: "com.github.hierynomus.license"
     apply plugin: 'eclipse'
     apply plugin: 'idea'
@@ -43,8 +43,8 @@ allprojects {
 subprojects {
     apply plugin: 'java'
 
-    sourceCompatibility = 11
-    targetCompatibility = 11
+    sourceCompatibility = 17
+    targetCompatibility = 17
     tasks.withType(JavaCompile) {
         options.encoding = 'UTF-8'
     }
@@ -53,20 +53,19 @@ subprojects {
     }
 
     dependencies {
-        compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.14.2'
-        compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.14.2'
-        compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.14.2'
-        compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-smile', version: '2.14.2'
-        compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-cbor', version: '2.14.2'
-        compile group: 'com.google.re2j', name: 're2j', version: '1.7'
-        testCompile group: 'org.powermock', name: 'powermock-module-javaagent', version: '2.0.9'
-        testCompile group: 'org.jacoco', name: 'org.jacoco.agent', version: '0.8.8', classifier: 'runtime'
-        testCompile group: 'org.powermock', name: 'powermock-core', version: '2.0.9'
+        implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.14.2'
+        implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.14.2'
+        implementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.14.2'
+        implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-smile', version: '2.14.2'
+        implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-cbor', version: '2.14.2'
+        implementation group: 'com.google.re2j', name: 're2j', version: '1.7'
+        testImplementation group: 'org.powermock', name: 'powermock-module-javaagent', version: '2.0.9'
+        testImplementation group: 'org.jacoco', name: 'org.jacoco.agent', version: '0.8.8', classifier: 'runtime'
+        testImplementation group: 'org.powermock', name: 'powermock-core', version: '2.0.9'
 
-        testCompile("junit:junit")
-        testCompile("org.springframework.boot:spring-boot-starter-test")
+        testImplementation("junit:junit")
+//        testImplementation("org.springframework.boot:spring-boot-starter-test")
     }
 
 
 }
-
